+++ examples/index.js
@@ -288,7 +288,7 @@ optionsTogglerEl.addEventListener('click', handleOptionsTogglerClick);
 createPiles(exampleEl.value).then(([pilingLib, additionalOptions = []]) => {
   const firstItem = pilingLib.get('items')[0];
 
-  const excludedProps = ['src', 'id', 'index'];
+  const excludedProps = ['src', 'id'];
   const numericalProps = Object.keys(firstItem).filter(
     prop =>
       excludedProps.indexOf(prop) === -1 && !Number.isNaN(+firstItem[prop])
@@ -735,6 +735,19 @@ createPiles(exampleEl.value).then(([pilingLib, additionalOptions = []]) => {
           name: 'Split All',
           action: () => {
             pilingLib.splitAll();
+          }
+        }
+      ]
+    },
+    {
+      id: 'split',
+      title: 'Split By',
+      fields: [
+        {
+          name: 'Split All',
+          width: '4rem',
+          action: () => {
+            piling.splitAll();
           }
         }
       ]